\name{affyGG-internal}

% Functions
\alias{dropprobes}

% Datasets
\alias{batch}
\alias{chrOffsets}
\alias{markersPos}
\alias{probeQtlProfiles}
\alias{probesetQtlProfiles}

\title{Internal affyGG functions}

\description{
  Internal affyGG functions. These are generally not to be called by the user.
}

\author{Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
